Got some stuff working.
So I've been staring at the spec for a few days now, trying to figure
out how to map types in the spec to C types and whatnot. That doesn't
seem to work. Best thing I could do right now is to make something
that works, even if it's fundamentally wrong, to avoid burnout and
solidify some intuitions. So, I did exactly that.